### 2021-12-17 #3999 @whedon
[PRE REVIEW]: SimpleFOC: A Field Oriented Control (FOC) Library for Controlling Brushless Direct Current (BLDC) and Stepper Motors
**Submitting author:** @askuric (<a href="http://orcid.org/0000-0002-3323-4482">Antun Skuric</a>)
**Repository:** https://github.com/simplefoc/Arduino-FOC/
**Branch with paper.md** (empty if default branch):
**Version:** v2.2
