### 2021-12-17 #3999 @kyleniemeyer
Hi @askuric, while we are getting started with assigning an editor and finding reviewers, could you take a look at the missing DOIs for references in your paper and begin adding those?
