### 2021-12-20 #3999 @askuric
Hello @kellyrowland,
Thanks for setting up the pre-review, I am very happy to be starting this process.
I've updated the references to include the DOIs and ISBNs where applicable.

Regarding the reviewers, I've passed through the list and I'd maybe suggest some experts in the control systems domain:
- sea-bass
- destogl
- ixjlyons
- CameronDevine
