+++ src/drivers/hardware_specific/esp32_mcu.cpp
@@ -1,6 +1,6 @@
 #include "../hardware_api.h"
 
-#if defined(ESP_H) && defined(ARDUINO_ARCH_ESP32)
+#if defined(ESP_H) && defined(ARDUINO_ARCH_ESP32) && defined(SOC_MCPWM_SUPPORTED) 
 
 #include "driver/mcpwm.h"
 #include "soc/mcpwm_reg.h"
