+++ src/drivers/hardware_specific/esp32_nomcpwm.cpp
@@ -1,6 +1,6 @@
 #include "../hardware_api.h"
 
-#if defined(ESP_H) && defined(ARDUINO_ARCH_ESP32) !defined(SOC_MCPWM_SUPPORTED)
+#if defined(ESP_H) && defined(ARDUINO_ARCH_ESP32) && !defined(SOC_MCPWM_SUPPORTED)
 
 #include "driver/ledc.h"
 
