### 2022-01-06 a3d72c0
inital navie support of the esp32s2 and esp32s3 - issue #120, #109
