### 2022-01-07 1ff4c06
Modify PWM timer configuration according to API changes in esp32 official SDK 2.0.2
