+++ src/drivers/hardware_specific/esp32_ledc_mcu.cpp
@@ -1,0 +1,204 @@
+#include "../hardware_api.h"
+
+#if defined(ESP_H) && defined(ARDUINO_ARCH_ESP32)// && !defined(SOC_MCPWM_SUPPORTED)
+
+#include "driver/ledc.h"
+
+#define _PWM_FREQUENCY 25000 // 25khz
+#define _PWM_FREQUENCY_MAX 38000 // 38khz max to be able to have 10 bit pwm resolution
+#define _PWM_RES_BIT 10  // 10 bir resolution
+#define _PWM_RES 1023 // 2^10-1 = 1023-1
+
+
+// empty motor slot
+#define _EMPTY_SLOT -20
+#define _TAKEN_SLOT -21
+
+// figure out how many ledc channels are avaible
+// esp32   - 2x8=16
+// esp32s2 - 8
+// esp32c3 - 6
+#include "soc/soc_caps.h"
+#ifdef SOC_LEDC_SUPPORT_HS_MODE
+#define LEDC_CHANNELS           (SOC_LEDC_CHANNEL_NUM<<1)
+#else
+#define LEDC_CHANNELS           (SOC_LEDC_CHANNEL_NUM)
+#endif
+
+// current channel stack index
+// support for multiple motors
+// esp32 has 16 channels
+// esp32s2 has 8 channels
+// esp32c3 has 6 channels
+int channel_index = 0;
+
+// slot for the 3pwm bldc motors
+typedef struct {
+  int pinID;
+  int ch1;
+  int ch2;
+  int ch3;
+} bldc_3pwm_motor_slots_t;
+
+// slot for the 2pwm stepper motors
+typedef struct {
+  int pinID;
+  int ch1;
+  int ch2;
+} stepper_2pwm_motor_slots_t;
+
+// slot for the 4pwm stepper motors
+typedef struct {
+  int pinID;
+  int ch1;
+  int ch2;
+  int ch3;
+  int ch4;
+} stepper_4pwm_motor_slots_t;
+
+
+// define motor slots array
+bldc_3pwm_motor_slots_t esp32_bldc_3pwm_motor_slots[4] = {
+  {_EMPTY_SLOT, 0,0,0}, // 1st motor
+  {_EMPTY_SLOT, 0,0,0}, // 2nd motor
+  {_EMPTY_SLOT, 0,0,0}, // 3st motor // esp32s2 & esp32
+  {_EMPTY_SLOT, 0,0,0}, // 4nd motor // esp32 only
+};
+stepper_2pwm_motor_slots_t esp32_stepper_2pwm_motor_slots[4]={
+  {_EMPTY_SLOT, 0,0}, // 1st motor
+  {_EMPTY_SLOT, 0,0}, // 2nd motor
+  {_EMPTY_SLOT, 0,0}, // 3rd motor
+  {_EMPTY_SLOT, 0,0}  // 4th motor - esp32s2 and esp32
+};
+stepper_4pwm_motor_slots_t esp32_stepper_4pwm_motor_slots[4]={
+  {_EMPTY_SLOT, 0,0,0,0}, // 1st motor
+  {_EMPTY_SLOT, 0,0,0,0}, // 2nd motor - esp32s2 and esp32
+  {_EMPTY_SLOT, 0,0,0,0}, // 3st motor - only esp32
+  {_EMPTY_SLOT, 0,0,0,0}, // 4st motor - only esp32
+};
+
+//  configure High PWM frequency
+void _setHighFrequency(const long freq, const int pin, const int channel){
+  ledcSetup(channel, freq, _PWM_RES_BIT );
+  ledcAttachPin(pin, channel);
+}
+
[…]
