### 2022-01-07 3816c82
workflow update for esp32 package v2.0.2 and esp32s2
