+++ .github/workflows/ccpp.yml
@@ -13,6 +13,7 @@ jobs:
           - arduino:samd:nano_33_iot                 # samd21
           - adafruit:samd:adafruit_metro_m4          # samd51
           - esp32:esp32:esp32doit-devkit-v1          # esm32
+          - esp32:esp32:esp32s2                      # esm32s2
           - STM32:stm32:GenF1:pnum=BLUEPILL_F103C8   # stm32 bluepill
           - STM32:stm32:Nucleo_64:pnum=NUCLEO_F411RE # stm32 nucleo
           - arduino:mbed_rp2040:pico                 # rpi pico
@@ -36,15 +37,19 @@ jobs:
             platform-url: https://adafruit.github.io/arduino-board-index/package_adafruit_index.json
             sketch-names: single_full_control_example.ino
           
+          - arduino-boards-fqbn: esp32:esp32:esp32s2 # esp32s2
+            platform-url: https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_dev_index.json
+            sketch-names: bldc_driver_3pwm_standalone.ino, stepper_driver_2pwm_standalone.ino, stepper_driver_4pwm_standalone 
+            
           - arduino-boards-fqbn: esp32:esp32:esp32doit-devkit-v1 # esp32 
-            platform-url: https://dl.espressif.com/dl/package_esp32_index.json
+            platform-url: https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_dev_index.json
             sketch-names: esp32_position_control.ino, esp32_i2c_dual_bus_example.ino, esp32_current_control_low_side.ino, esp32_spi_alt_example.ino
 
           - arduino-boards-fqbn: STM32:stm32:GenF1:pnum=BLUEPILL_F103C8 # bluepill - hs examples
             platform-url: https://github.com/stm32duino/BoardManagerFiles/raw/master/STM32/package_stm_index.json 
             sketch-names: bluepill_position_control.ino, stm32_i2c_dual_bus_example.ino, stm32_spi_alt_example.ino
             
-          - arduino-boards-fqbn: STM32:stm32:Nucleo_64:pnum=NUCLEO_F411RE # one full example
+          - arduino-boards-fqbn: STM32:stm32:Nucleo_64:pnum=NUCLEO_F411RE # nucleo one full example
             platform-url: https://github.com/stm32duino/BoardManagerFiles/raw/master/STM32/package_stm_index.json 
             sketch-names: single_full_control_example.ino, stm32_spi_alt_example.ino
             
