### 2022-01-07 e87dd58
update workflow + refactoring
