+++ .github/workflows/ccpp.yml
@@ -12,8 +12,8 @@ jobs:
           - arduino:sam:arduino_due_x                # arduino due
           - arduino:samd:nano_33_iot                 # samd21
           - adafruit:samd:adafruit_metro_m4          # samd51
-          - esp32:esp32:esp32doit-devkit-v1          # esm32
-          - esp32:esp32:esp32s2                      # esm32s2
+          - esp32:esp32:esp32                        # esp32
+          - esp32:esp32:esp32s2                      # esp32s2
           - STM32:stm32:GenF1:pnum=BLUEPILL_F103C8   # stm32 bluepill
           - STM32:stm32:Nucleo_64:pnum=NUCLEO_F411RE # stm32 nucleo
           - arduino:mbed_rp2040:pico                 # rpi pico
@@ -41,7 +41,7 @@ jobs:
             platform-url: https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_dev_index.json
             sketch-names: bldc_driver_3pwm_standalone.ino, stepper_driver_2pwm_standalone.ino, stepper_driver_4pwm_standalone 
             
-          - arduino-boards-fqbn: esp32:esp32:esp32doit-devkit-v1 # esp32 
+          - arduino-boards-fqbn: esp32:esp32:esp32 # esp32 
             platform-url: https://raw.githubusercontent.com/espressif/arduino-esp32/gh-pages/package_esp32_dev_index.json
             sketch-names: esp32_position_control.ino, esp32_i2c_dual_bus_example.ino, esp32_current_control_low_side.ino, esp32_spi_alt_example.ino
 
