+++ README.md
@@ -26,6 +26,7 @@ Therefore this is an attempt to:
       <li>Voltage/current limit handling bugs #118</li>
       <li>Generic position and current sense classes - to implement a new sensor only implement one function</li>
       <li>Initial support for esp32s2 and esp32s3 - separation of the esp32 mcpwm and led implementation</li>
+      <li><b>esp32 arduino package transfer to v2.0.1+</b></li>
    </ul>
 </blockquote>
 
