### 2022-01-16 #3999 @gkthiruvathukal
@askuric Thank you for your patience. I was out of the office until 4th January. Then I became department chair at my university and have been transitioning into this role. I'm finally able to resume work on my JOSS workload.

@sea-bass and @destogl, Are you willing to contribute a review for this JOSS submission?
