+++ src/drivers/hardware_specific/samd_mcu.h
@@ -80,6 +80,14 @@ struct wo_association {
 
 
 
+struct SAMDHardwareDriverParams : public HardwareDriverParamsBase {
+	SAMDHardwareDriverParams(int numPins) : tccPinConfigurations[numPins] {};
+	tccConfiguration tccPinConfigurations[6];
+};
+
+
+
+
 #if defined(_SAMD21_)
 #define NUM_PIO_TIMER_PERIPHERALS 2
 #elif defined(_SAMD51_)||defined(_SAME51_)
