+++ src/drivers/hardware_specific/samd_mcu.cpp
@@ -278,7 +278,7 @@ int checkPermutations(uint8_t num, int pins[], bool (*checkFunc)(tccConfiguratio
  * @param pinA pinA bldc driver
  * @param pinB pinB bldc driver
  */
-void _configure2PWM(long pwm_frequency, const int pinA, const int pinB) {
+HardwareDriverParams _configure2PWM(long pwm_frequency, const int pinA, const int pinB) {
 #ifdef SIMPLEFOC_SAMD_DEBUG
 	printAllPinInfos();
 #endif
@@ -375,7 +375,7 @@ void _configure2PWM(long pwm_frequency, const int pinA, const int pinB) {
  * @param pinB pinB bldc driver
  * @param pinC pinC bldc driver
  */
-void _configure3PWM(long pwm_frequency, const int pinA, const int pinB, const int pinC) {
+HardwareDriverParams _configure3PWM(long pwm_frequency, const int pinA, const int pinB, const int pinC) {
 #ifdef SIMPLEFOC_SAMD_DEBUG
 	printAllPinInfos();
 #endif
@@ -451,7 +451,7 @@ void _configure3PWM(long pwm_frequency, const int pinA, const int pinB, const in
  * @param pin2A pin2A stepper driver
  * @param pin2B pin2B stepper driver
  */
-void _configure4PWM(long pwm_frequency, const int pin1A, const int pin1B, const int pin2A, const int pin2B) {
+HardwareDriverParams _configure4PWM(long pwm_frequency, const int pin1A, const int pin1B, const int pin2A, const int pin2B) {
 #ifdef SIMPLEFOC_SAMD_DEBUG
 	printAllPinInfos();
 #endif
@@ -552,7 +552,7 @@ void _configure4PWM(long pwm_frequency, const int pin1A, const int pin1B, const 
  *
  * @return 0 if config good, -1 if failed
  */
-int _configure6PWM(long pwm_frequency, float dead_zone, const int pinA_h, const int pinA_l,  const int pinB_h, const int pinB_l, const int pinC_h, const int pinC_l) {
+HardwareDriverParams _configure6PWM(long pwm_frequency, float dead_zone, const int pinA_h, const int pinA_l,  const int pinB_h, const int pinB_l, const int pinC_h, const int pinC_l) {
 	// we want to use a TCC channel with 1 non-inverted and 1 inverted output for each phase, with dead-time insertion
 #ifdef SIMPLEFOC_SAMD_DEBUG
 	printAllPinInfos();
@@ -644,7 +644,7 @@ int _configure6PWM(long pwm_frequency, float dead_zone, const int pinA_h, const 
  * @param pinA  phase A hardware pin number
  * @param pinB  phase B hardware pin number
  */
-void _writeDutyCycle2PWM(float dc_a,  float dc_b, int pinA, int pinB) {
+void _writeDutyCycle2PWM(float dc_a,  float dc_b, HardwareDriverParams params) {
 	tccConfiguration* tccI = getTccPinConfiguration(pinA);
 	writeSAMDDutyCycle(tccI, dc_a);
 	tccI = getTccPinConfiguration(pinB);
@@ -668,7 +668,7 @@ void _writeDutyCycle2PWM(float dc_a,  float dc_b, int pinA, int pinB) {
  * @param pinB  phase B hardware pin number
  * @param pinC  phase C hardware pin number
  */
-void _writeDutyCycle3PWM(float dc_a,  float dc_b, float dc_c, int pinA, int pinB, int pinC) {
[…]
