### 2022-02-01 7f02b8a
driver API refactor, part 1, the easy ones
