+++ src/drivers/hardware_api.h
@@ -4,16 +4,13 @@
 #include "../common/foc_utils.h"
 #include "../common/time_utils.h"
 
+#define SIMPLEFOC_DRIVER_INIT_FAILED ((void*)-1)
 
-struct HardwareDriverParamsBase {
-    int pins[6];
-    long pwm_frequency;
-    int dead_zone;
-    bool initSuccess;
-};
-
-typedef struct HardwareDriverParamsBase* HardwareDriverParams;
-
+typedef struct GenericDriverParams {
+  int pins[6];
+  long pwm_frequency;
+  float dead_zone;
+} GenericDriverParams;
 
 
 /** 
@@ -24,8 +21,10 @@ typedef struct HardwareDriverParamsBase* HardwareDriverParams;
  * @param pwm_frequency - frequency in hertz - if applicable
  * @param pinA pinA bldc driver
  * @param pinB pinB bldc driver
+ * 
+ * @return -1 if failed, or pointer to internal driver parameters struct if successful
  */
-HardwareDriverParams _configure2PWM(long pwm_frequency, const int pinA, const int pinB);
+void* _configure2PWM(long pwm_frequency, const int pinA, const int pinB);
 
 /** 
  * Configuring PWM frequency, resolution and alignment
@@ -36,8 +35,10 @@ HardwareDriverParams _configure2PWM(long pwm_frequency, const int pinA, const in
  * @param pinA pinA bldc driver
  * @param pinB pinB bldc driver
  * @param pinC pinC bldc driver
+ * 
+ * @return -1 if failed, or pointer to internal driver parameters struct if successful
  */
-HardwareDriverParams _configure3PWM(long pwm_frequency, const int pinA, const int pinB, const int pinC);
+void* _configure3PWM(long pwm_frequency, const int pinA, const int pinB, const int pinC);
 
 /** 
  * Configuring PWM frequency, resolution and alignment
@@ -49,8 +50,10 @@ HardwareDriverParams _configure3PWM(long pwm_frequency, const int pinA, const in
  * @param pin1B pin1B stepper driver
  * @param pin2A pin2A stepper driver
  * @param pin2B pin2B stepper driver
+ * 
+ * @return -1 if failed, or pointer to internal driver parameters struct if successful
  */
-HardwareDriverParams _configure4PWM(long pwm_frequency, const int pin1A, const int pin1B, const int pin2A, const int pin2B);
+void* _configure4PWM(long pwm_frequency, const int pin1A, const int pin1B, const int pin2A, const int pin2B);
 
 /** 
  * Configuring PWM frequency, resolution and alignment
@@ -66,9 +69,9 @@ HardwareDriverParams _configure4PWM(long pwm_frequency, const int pin1A, const i
  * @param pinC_h pinA high-side bldc driver 
  * @param pinC_l pinA low-side bldc driver 
  * 
- * @return 0 if config good, -1 if failed
+ * @return -1 if failed, or pointer to internal driver parameters struct if successful
  */
-HardwareDriverParams _configure6PWM(long pwm_frequency, float dead_zone, const int pinA_h, const int pinA_l,  const int pinB_h, const int pinB_l, const int pinC_h, const int pinC_l);
+void* _configure6PWM(long pwm_frequency, float dead_zone, const int pinA_h, const int pinA_l,  const int pinB_h, const int pinB_l, const int pinC_h, const int pinC_l);
 
 /** 
  * Function setting the duty cycle to the pwm pin (ex. analogWrite())
[…]
