+++ src/drivers/hardware_specific/samd_mcu.cpp
@@ -278,7 +278,7 @@ int checkPermutations(uint8_t num, int pins[], bool (*checkFunc)(tccConfiguratio
  * @param pinA pinA bldc driver
  * @param pinB pinB bldc driver
  */
-HardwareDriverParams _configure2PWM(long pwm_frequency, const int pinA, const int pinB) {
+void* _configure2PWM(long pwm_frequency, const int pinA, const int pinB) {
 #ifdef SIMPLEFOC_SAMD_DEBUG
 	printAllPinInfos();
 #endif
@@ -289,7 +289,7 @@ HardwareDriverParams _configure2PWM(long pwm_frequency, const int pinA, const in
 #ifdef SIMPLEFOC_SAMD_DEBUG
 		SIMPLEFOC_SAMD_DEBUG_SERIAL.println("Bad combination!");
 #endif
-		return;
+		return SIMPLEFOC_DRIVER_INIT_FAILED;
 	}
 
 	tccConfiguration tccConfs[2] = { getTCCChannelNr(pinA, getPeripheralOfPermutation(compatibility, 0)),
@@ -329,7 +329,11 @@ HardwareDriverParams _configure2PWM(long pwm_frequency, const int pinA, const in
 	SIMPLEFOC_SAMD_DEBUG_SERIAL.println("Configured TCCs...");
 #endif
 
-	return; // Someone with a stepper-setup who can test it?
+	SAMDHardwareDriverParams* params = new SAMDHardwareDriverParams {
+		.tccPinConfigurations = { getTccPinConfiguration(pinA), getTccPinConfiguration(pinB) },
+		.pwm_frequency = (uint32_t)pwm_frequency
+	}; // Someone with a stepper-setup who can test it?
+	return params;
 }
 
 
@@ -375,7 +379,7 @@ HardwareDriverParams _configure2PWM(long pwm_frequency, const int pinA, const in
  * @param pinB pinB bldc driver
  * @param pinC pinC bldc driver
  */
-HardwareDriverParams _configure3PWM(long pwm_frequency, const int pinA, const int pinB, const int pinC) {
+void* _configure3PWM(long pwm_frequency, const int pinA, const int pinB, const int pinC) {
 #ifdef SIMPLEFOC_SAMD_DEBUG
 	printAllPinInfos();
 #endif
@@ -386,7 +390,7 @@ HardwareDriverParams _configure3PWM(long pwm_frequency, const int pinA, const in
 #ifdef SIMPLEFOC_SAMD_DEBUG
 		SIMPLEFOC_SAMD_DEBUG_SERIAL.println("Bad combination!");
 #endif
-		return;
+		return SIMPLEFOC_DRIVER_INIT_FAILED;
 	}
 
 	tccConfiguration tccConfs[3] = { getTCCChannelNr(pinA, getPeripheralOfPermutation(compatibility, 0)),
@@ -431,6 +435,11 @@ HardwareDriverParams _configure3PWM(long pwm_frequency, const int pinA, const in
 	SIMPLEFOC_SAMD_DEBUG_SERIAL.println("Configured TCCs...");
 #endif
 
+	return new SAMDHardwareDriverParams {
+		.tccPinConfigurations = { getTccPinConfiguration(pinA), getTccPinConfiguration(pinB), getTccPinConfiguration(pinC) },
+		.pwm_frequency = (uint32_t)pwm_frequency
+	};
+
 }
 
 
@@ -451,7 +460,7 @@ HardwareDriverParams _configure3PWM(long pwm_frequency, const int pinA, const in
  * @param pin2A pin2A stepper driver
[…]
