+++ src/drivers/StepperDriver2PWM.cpp
@@ -80,7 +80,7 @@ int StepperDriver2PWM::init() {
   // Set the pwm frequency to the pins
   // hardware specific function - depending on driver and mcu
   params = _configure2PWM(pwm_frequency, pwm1, pwm2);
-  return params!=0 && params->initSuccess;
+  return params!=SIMPLEFOC_DRIVER_INIT_FAILED;
 }
 
 
