+++ src/drivers/StepperDriver4PWM.cpp
@@ -55,7 +55,7 @@ int StepperDriver4PWM::init() {
   // Set the pwm frequency to the pins
   // hardware specific function - depending on driver and mcu
   params = _configure4PWM(pwm_frequency, pwm1A, pwm1B, pwm2A, pwm2B);
-  return params!=0 && params->initSuccess;
+  return params!=SIMPLEFOC_DRIVER_INIT_FAILED;
 }
 
 

[…]
