### 2022-02-03 84269cd
commander refactorred and separated motion control into a new callback
