+++ src/communication/Commander.cpp
@@ -102,7 +102,7 @@ void Commander::motor(FOCMotor* motor, char* user_command) {
   // if target setting
   if(isDigit(user_command[0]) || user_command[0] == '-' || user_command[0] == '+'){
     printVerbose(F("Target: "));
-    motor->target = atof(user_command);
+    target(motor, user_command);
     println(motor->target);
     return;
   }
@@ -177,65 +177,9 @@ void Commander::motor(FOCMotor* motor, char* user_command) {
       }
       break;
     case CMD_MOTION_TYPE:
-      printVerbose(F("Motion:"));
-      switch(sub_cmd){
-        case SCMD_DOWNSAMPLE:
-            printVerbose(F(" downsample: "));
-            if(!GET) motor->motion_downsample = value;
-            println((int)motor->motion_downsample);
-          break;
-        default:
-          // change control type
-          if(!GET && value >= 0 && (int)value < 5) // if set command
-            motor->controller = (MotionControlType)value;
-          switch(motor->controller){
-            case MotionControlType::torque:
-              println(F("torque"));
-              break;
-            case MotionControlType::velocity:
-              println(F("vel"));
-              break;
-            case MotionControlType::angle:
-              println(F("angle"));
-              break;
-            case MotionControlType::velocity_openloop:
-              println(F("vel open"));
-              break;
-            case MotionControlType::angle_openloop:
-              println(F("angle open"));
-              break;
-          }
-            break;
-        }
-      break;
     case CMD_TORQUE_TYPE:
-      // change control type
-      printVerbose(F("Torque: "));
-      if(!GET && (int8_t)value >= 0 && (int8_t)value < 3)// if set command
-        motor->torque_controller = (TorqueControlType)value;
-      switch(motor->torque_controller){
-        case TorqueControlType::voltage:
-          println(F("volt"));
-          // change the velocity control limits if necessary
-          if( !_isset(motor->phase_resistance) ) motor->PID_velocity.limit = motor->voltage_limit;
-          break;
-        case TorqueControlType::dc_current:
-          println(F("dc curr"));
-          // change the velocity control limits if necessary
-          motor->PID_velocity.limit = motor->current_limit;
-          break;
-        case TorqueControlType::foc_current:
-          println(F("foc curr"));
-          // change the velocity control limits if necessary
-          motor->PID_velocity.limit = motor->current_limit;
-          break;
-      }
-      break;
     case CMD_STATUS:
-      // enable/disable
-      printVerbose(F("Status: "));
-      if(!GET) (bool)value ? motor->enable() : motor->disable();
-       println(motor->enabled);
+      motion(motor, &user_command[0]);
       break;
     case CMD_PWMMOD:
        // PWM modulation change
@@ -384,6 +328,80 @@ void Commander::motor(FOCMotor* motor, char* user_command) {
   }
 }
 
+void Commander::motion(FOCMotor* motor, char* user_cmd, char* separator){
+  char cmd = user_cmd[0];
+  char sub_cmd = user_cmd[1];
+  bool GET  = isSentinel(user_cmd[1]);
+  float value = atof(&user_cmd[(sub_cmd >= 'A'  && sub_cmd <= 'Z') ?  2 :  1]);
+
+  switch(cmd){
+    case CMD_MOTION_TYPE:
+      printVerbose(F("Motion:"));
+      switch(sub_cmd){
+        case SCMD_DOWNSAMPLE:
+            printVerbose(F(" downsample: "));
+            if(!GET) motor->motion_downsample = value;
+            println((int)motor->motion_downsample);
+          break;
+        default:
+          // change control type
+          if(!GET && value >= 0 && (int)value < 5) // if set command
+            motor->controller = (MotionControlType)value;
+          switch(motor->controller){
+            case MotionControlType::torque:
+              println(F("torque"));
[…]
