+++ src/communication/Commander.h
@@ -100,9 +100,13 @@ class Commander
     Stream* com_port = nullptr; //!< Serial terminal variable if provided
     char eol = '\n'; //!< end of line sentinel character
     bool echo = false; //!< echo last typed character (for command line feedback)
+
     /**
      *
      * FOC motor (StepperMotor and BLDCMotor) command interface
+     * @param motor    - FOCMotor (BLDCMotor or StepperMotor) instance 
+     * @param user_cmd - the string command
+     * 
      *  - It has several paramters (the letters can be changed in the commands.h file)
      *    'Q' - Q current PID controller & LPF (see function pid and lpf for commands)
      *    'D' - D current PID controller & LPF (see function pid and lpf for commands)
@@ -149,6 +153,9 @@ class Commander
 
     /**
      * Low pass fileter command interface
+     * @param lpf      - LowPassFilter instance 
+     * @param user_cmd - the string command
+     * 
      *  - It only has one property - filtering time constant Tf
      *  - It can be get by sending 'F'
      *  - It can be set by sending 'Fvalue' - (ex. F0.01 for settin Tf=0.01)
@@ -156,6 +163,9 @@ class Commander
     void lpf(LowPassFilter* lpf, char* user_cmd);
     /**
      * PID controller command interface
+     * @param pid      - PIDController instance 
+     * @param user_cmd - the string command
+     * 
      *  - It has several paramters (the letters can be changed in the commands.h file)
      *     - P gain       - 'P'
      *     - I gain       - 'I'
@@ -168,6 +178,9 @@ class Commander
     void pid(PIDController* pid, char* user_cmd);
     /**
      * Float variable scalar command interface
+     * @param value    - float variable pointer 
+     * @param user_cmd - the string command
+     * 
      *  - It only has one property - one float value
      *  - It can be get by sending an empty string '\n'
      *  - It can be set by sending 'value' - (ex. 0.01f for settin *value=0.01)
@@ -175,15 +188,53 @@ class Commander
     void scalar(float* value, char* user_cmd);
     /**
      *  Target setting interface, enables setting the target and limiting variables at once. 
-     *  The valeus are sent separated by a space. ex. P2.34 70 2
+     *  The values are sent separated by a separator specified as the third argument. The default separator is the space.
+     * 
+     * @param motor     - FOCMotor (BLDCMotor or StepperMotor) instance 
+     * @param user_cmd  - the string command
+     * @param separator - the string separator in between target and limit values, default is space - " "
+     *  
+     *  Example: P2.34 70 2
      *  `P` is the user defined command, `2.34` is the target angle `70` is the target 
      *  velocity and `2` is the desired max current.
+     *  
      *  It depends of the motion control mode:
-     *  - torque   : torque (ex. P2.5) 
-     *  - velocity : velocity torque (ex.P10 2.5) 
-     *  - angle    : angle velocity torque (ex.P3.5 10 2.5)
+     *    - torque   : torque (ex. P2.5) 
+     *    - velocity : velocity torque (ex.P10 2.5 or P10 to only chanage the target witout limits)
+     *    - angle    : angle velocity torque (ex.P3.5 10 2.5 or P3.5 to only chanage the target witout limits)
      */
-    void target(FOCMotor* motor, char* user_cmd);
+    void target(FOCMotor* motor, char* user_cmd, char* separator = " ");
+
+    /**
+     * FOC motor (StepperMotor and BLDCMotor) motion control interfaces
+     * @param motor     - FOCMotor (BLDCMotor or StepperMotor) instance 
+     * @param user_cmd  - the string command
+     * @param separator - the string separator in between target and limit values, default is space - " "
+     * 
+     * Commands:
+     *    'C' - Motion control type config
[…]
