### 2022-02-05 fba8876
STM32 refactored (but untested)
