+++ src/drivers/hardware_specific/stm32_mcu.cpp
@@ -14,15 +14,28 @@
 #define _ERROR_6PWM -1
 
 
-// setting pwm to hardware pin - instead analogWrite()
-void _setPwm(int ulPin, uint32_t value, int resolution)
-{
-  PinName pin = digitalPinToPinName(ulPin);
-  TIM_TypeDef *Instance = (TIM_TypeDef *)pinmap_peripheral(pin, PinMap_PWM);
-  uint32_t index = get_timer_index(Instance);
-  HardwareTimer *HT = (HardwareTimer *)(HardwareTimer_Handle[index]->__this);
 
-  uint32_t channel = STM_PIN_CHANNEL(pinmap_function(pin, PinMap_PWM));
+typedef struct STM32DriverParams {
+  HardwareTimer* timers[6];
+  uint32_t channels[6];
+  long pwm_frequency;
+  float dead_zone;
+  uint8_t interface_type;
+} STM32DriverParams;
+
+
+
+
+
+// setting pwm to hardware pin - instead analogWrite()
+void _setPwm(HardwareTimer *HT, uint32_t channel, uint32_t value, int resolution)
+{
+  // PinName pin = digitalPinToPinName(ulPin);
+  // TIM_TypeDef *Instance = (TIM_TypeDef *)pinmap_peripheral(pin, PinMap_PWM);
+  // uint32_t index = get_timer_index(Instance);
+  // HardwareTimer *HT = (HardwareTimer *)(HardwareTimer_Handle[index]->__this);
+
+  //uint32_t channel = STM_PIN_CHANNEL(pinmap_function(pin, PinMap_PWM));
   HT->setCaptureCompare(channel, value, (TimerCompareFormat_t)resolution);
 }
 
@@ -129,7 +142,7 @@ void _alignPWMTimers(HardwareTimer *HT1,HardwareTimer *HT2,HardwareTimer *HT3,Ha
 }
 
 // configure hardware 6pwm interface only one timer with inverted channels
-HardwareTimer* _initHardware6PWMInterface(uint32_t PWM_freq, float dead_zone, int pinA_h, int pinA_l, int pinB_h, int pinB_l, int pinC_h, int pinC_l)
+STM32DriverParams* _initHardware6PWMInterface(uint32_t PWM_freq, float dead_zone, int pinA_h, int pinA_l, int pinB_h, int pinB_l, int pinC_h, int pinC_l)
 {
   
 #if !defined(STM32L0xx) // L0 boards dont have hardware 6pwm interface 
@@ -139,6 +152,12 @@ HardwareTimer* _initHardware6PWMInterface(uint32_t PWM_freq, float dead_zone, in
   PinName vlPinName = digitalPinToPinName(pinB_l);
   PinName whPinName = digitalPinToPinName(pinC_h);
   PinName wlPinName = digitalPinToPinName(pinC_l);
+  uint32_t channel1 = STM_PIN_CHANNEL(pinmap_function(uhPinName, PinMap_PWM));
+  uint32_t channel2 = STM_PIN_CHANNEL(pinmap_function(ulPinName, PinMap_PWM));
+  uint32_t channel3 = STM_PIN_CHANNEL(pinmap_function(vhPinName, PinMap_PWM));
+  uint32_t channel4 = STM_PIN_CHANNEL(pinmap_function(vlPinName, PinMap_PWM));
+  uint32_t channel5 = STM_PIN_CHANNEL(pinmap_function(whPinName, PinMap_PWM));
+  uint32_t channel6 = STM_PIN_CHANNEL(pinmap_function(wlPinName, PinMap_PWM));
 
[…]
