### 2022-02-05 514f1c1
STM32 driver refactor (not yet tested)
