+++ src/drivers/hardware_specific/esp32_mcu.cpp
@@ -35,6 +35,7 @@ typedef struct {
   mcpwm_io_signals_t mcpwm_b;
   mcpwm_io_signals_t mcpwm_c;
 } bldc_3pwm_motor_slots_t;
+
 typedef struct {
   int pin1A;
   mcpwm_dev_t* mcpwm_num;
@@ -46,6 +47,7 @@ typedef struct {
   mcpwm_io_signals_t mcpwm_2a;
   mcpwm_io_signals_t mcpwm_2b;
 } stepper_4pwm_motor_slots_t;
+
 typedef struct {
   int pin1pwm;
   mcpwm_dev_t* mcpwm_num;
@@ -68,6 +70,8 @@ typedef struct {
   mcpwm_io_signals_t mcpwm_bl;
   mcpwm_io_signals_t mcpwm_cl;
 } bldc_6pwm_motor_slots_t;
+
+
 
 // define bldc motor slots array
 bldc_3pwm_motor_slots_t esp32_bldc_3pwm_motor_slots[4] =  {
@@ -96,6 +100,19 @@ stepper_2pwm_motor_slots_t esp32_stepper_2pwm_motor_slots[4] =  {
   {_EMPTY_SLOT, &MCPWM1, MCPWM_UNIT_1, MCPWM_OPR_A, MCPWM0A, MCPWM1A}, // 3rd motor will be MCPWM1 channel A
   {_EMPTY_SLOT, &MCPWM1, MCPWM_UNIT_1, MCPWM_OPR_B, MCPWM0B, MCPWM1B}  // 4th motor will be MCPWM1 channel B
   };
+
+
+
+typedef struct ESP32MCPWMDriverParams {
+  long pwm_frequency;
+  mcpwm_unit_t mcpwm_unit;
+  mcpwm_operator_t mcpwm_operator1;
+  mcpwm_operator_t mcpwm_operator2;
+} ESP32MCPWMDriverParams;
+
+
+
+
 
 // configuring high frequency pwm timer
 // a lot of help from this post from Paul Gauld
@@ -180,7 +197,7 @@ void _configureTimerFrequency(long pwm_frequency, mcpwm_dev_t* mcpwm_num,  mcpwm
 // - Stepper motor - 2PWM setting
 // - hardware speciffic
 // supports Arudino/ATmega328, STM32 and ESP32
-void _configure2PWM(long pwm_frequency,const int pinA, const int pinB) {
+void* _configure2PWM(long pwm_frequency, const int pinA, const int pinB) {
   if(!pwm_frequency || !_isset(pwm_frequency) ) pwm_frequency = _PWM_FREQUENCY; // default frequency 25hz
   else pwm_frequency = _constrain(pwm_frequency, 0, _PWM_FREQUENCY_MAX); // constrain to 40kHz max
 
@@ -218,6 +235,12 @@ void _configure2PWM(long pwm_frequency,const int pinA, const int pinB) {
   // configure the timer
   _configureTimerFrequency(pwm_frequency, m_slot.mcpwm_num,  m_slot.mcpwm_unit);
 
+  ESP32MCPWMDriverParams* params = new ESP32MCPWMDriverParams {
+    .pwm_frequency = pwm_frequency,
+    .mcpwm_unit = m_slot.mcpwm_unit,
+    .mcpwm_operator1 = m_slot.mcpwm_operator
+  };
+  return params;
 }
 
 
@@ -225,7 +248,7 @@ void _configure2PWM(long pwm_frequency,const int pinA, const int pinB) {
 // - BLDC motor - 3PWM setting
 // - hardware speciffic
 // supports Arudino/ATmega328, STM32 and ESP32
-void _configure3PWM(long pwm_frequency,const int pinA, const int pinB, const int pinC) {
[…]
