### 2022-02-06 b9257bc
nrf52 driver refactored (but untested)
