+++ src/drivers/hardware_specific/nrf52_mcu.cpp
@@ -22,7 +22,7 @@
 #define _TAKEN_SLOT (-0x55)
 
 int pwm_range;
-float dead_time;
+
 
 static NRF_PWM_Type* pwms[PWM_COUNT] = {
   NRF_PWM0,
@@ -72,7 +72,22 @@ stepper_motor_slots_t nrf52_stepper_motor_slots[4] =  {
 bldc_6pwm_motor_slots_t nrf52_bldc_6pwm_motor_slots[2] =  { 
   {_EMPTY_SLOT, pwms[0], pwms[1], {0,0,0,0,0,0,0,0}},// 1st motor will be on PWM0 & PWM1
   {_EMPTY_SLOT, pwms[2], pwms[3], {0,0,0,0,0,0,0,0}} // 2nd motor will be on PWM1 & PWM2
-  };  
+  };
+
+
+
+typedef struct NRF52DriverParams {
+  union {
+    bldc_3pwm_motor_slots_t* slot3pwm;
+    bldc_6pwm_motor_slots_t* slot6pwm;
+    stepper_motor_slots_t* slotstep;
+  } slot;
+  long pwm_frequency;
+  float dead_time;
+} NRF52DriverParams;
+
+
+
 
 // configuring high frequency pwm timer
 void _configureHwPwm(NRF_PWM_Type* mcpwm1,  NRF_PWM_Type* mcpwm2){
@@ -103,7 +118,7 @@ void _configureHwPwm(NRF_PWM_Type* mcpwm1,  NRF_PWM_Type* mcpwm2){
 // function setting the high pwm frequency to the supplied pins
 // - BLDC motor - 3PWM setting
 // - hardware speciffic
-void _configure3PWM(long pwm_frequency,const int pinA, const int pinB, const int pinC) {
+void* _configure3PWM(long pwm_frequency,const int pinA, const int pinB, const int pinC) {
 
   if( !pwm_frequency || pwm_frequency == NOT_SET) pwm_frequency = PWM_FREQ; // default frequency 20khz for a resolution of 800
   else pwm_frequency = _constrain(pwm_frequency, 0, PWM_MAX_FREQ); // constrain to 62.5kHz max for a resolution of 256  
@@ -123,6 +138,9 @@ void _configure3PWM(long pwm_frequency,const int pinA, const int pinB, const int
       break;
     }
   }
+  // if no slots available
+  if(slot_num >= 4) return SIMPLEFOC_DRIVER_INIT_FAILED;
+
   // disable all the slots with the same MCPWM 
   if(slot_num < 2){
     // slot 0 of the stepper
@@ -149,12 +167,20 @@ void _configure3PWM(long pwm_frequency,const int pinA, const int pinB, const int
 
   // configure the pwm
   _configureHwPwm(nrf52_bldc_3pwm_motor_slots[slot_num].mcpwm, nrf52_bldc_3pwm_motor_slots[slot_num].mcpwm);
+
+  NRF52DriverParams* params = new NRF52DriverParams();
+  params->slot.slot3pwm = &(nrf52_bldc_3pwm_motor_slots[slot_num]);
+  params->pwm_frequency = pwm_frequency;
+  return params;
 }
+
+
+
 
 // function setting the high pwm frequency to the supplied pins
 // - Stepper motor - 4PWM setting
 // - hardware speciffic
-void _configure4PWM(long pwm_frequency,const int pinA, const int pinB, const int pinC, const int pinD) {
+void* _configure4PWM(long pwm_frequency, const int pinA, const int pinB, const int pinC, const int pinD) {
 
[…]
