### 2022-02-06 a2344ef
make 2-pwm unsupported explicit
