### 2022-02-07 8db941a
fixup SAMD driver refactor (working, tested)
