### 2022-02-12 5889c49
some comments and small typo for atmega328
