+++ src/BLDCMotor.cpp
@@ -24,7 +24,7 @@ void BLDCMotor::linkDriver(BLDCDriver* _driver) {
 
 // init hardware pins
 void BLDCMotor::init() {
-  if(monitor_port) monitor_port->println(F("MOT: Init"));
+  SIMPLEFOC_DEBUG("MOT: Init");
 
   // if no current sensing and the user has set the phase resistance of the motor use current limit to calculate the voltage limit
   if( !current_sense && _isset(phase_resistance)) {
@@ -54,7 +54,7 @@ void BLDCMotor::init() {
 
   _delay(500);
   // enable motor
-  if(monitor_port) monitor_port->println(F("MOT: Enable driver."));
+  SIMPLEFOC_DEBUG("MOT: Enable driver.");
   enable();
   _delay(500);
 }
@@ -104,7 +104,8 @@ int  BLDCMotor::initFOC( float zero_electric_offset, Direction _sensor_direction
     // added the shaft_angle update
     sensor->update();
     shaft_angle = shaftAngle();
-  }else if(monitor_port) monitor_port->println(F("MOT: No sensor."));
+  }else 
+    SIMPLEFOC_DEBUG("MOT: No sensor.");
 
   // aligning the current sensor - can be skipped
   // checks if driver phases are the same as current sense phases
@@ -112,13 +113,13 @@ int  BLDCMotor::initFOC( float zero_electric_offset, Direction _sensor_direction
   _delay(500);
   if(exit_flag){
     if(current_sense) exit_flag *= alignCurrentSense();
-    else if(monitor_port) monitor_port->println(F("MOT: No current sense."));
+    else SIMPLEFOC_DEBUG("MOT: No current sense.");
   }
 
   if(exit_flag){
-    if(monitor_port) monitor_port->println(F("MOT: Ready."));
+    SIMPLEFOC_DEBUG("MOT: Ready.");
   }else{
-    if(monitor_port) monitor_port->println(F("MOT: Init FOC failed."));
+    SIMPLEFOC_DEBUG("MOT: Init FOC failed.");
     disable();
   }
 
@@ -129,18 +130,17 @@ int  BLDCMotor::initFOC( float zero_electric_offset, Direction _sensor_direction
 int BLDCMotor::alignCurrentSense() {
   int exit_flag = 1; // success
 
-  if(monitor_port) monitor_port->println(F("MOT: Align current sense."));
+  SIMPLEFOC_DEBUG("MOT: Align current sense.");
 
   // align current sense and the driver
   exit_flag = current_sense->driverAlign(driver, voltage_sensor_align);
   if(!exit_flag){
     // error in current sense - phase either not measured or bad connection
-    if(monitor_port) monitor_port->println(F("MOT: Align error!"));
+    SIMPLEFOC_DEBUG("MOT: Align error!");
     exit_flag = 0;
   }else{
     // output the alignment status flag
-    if(monitor_port) monitor_port->print(F("MOT: Success: "));
-    if(monitor_port) monitor_port->println(exit_flag);
+    SIMPLEFOC_DEBUG("MOT: Success: ", exit_flag);
   }
 
   return exit_flag > 0;
@@ -149,7 +149,7 @@ int BLDCMotor::alignCurrentSense() {
 // Encoder alignment to electrical 0 angle
 int BLDCMotor::alignSensor() {
   int exit_flag = 1; //success
-  if(monitor_port) monitor_port->println(F("MOT: Align sensor."));
+  SIMPLEFOC_DEBUG("MOT: Align sensor.");
 
   // if unknown natural direction
   if(!_isset(sensor_direction)){
@@ -180,24 +180,23 @@ int BLDCMotor::alignSensor() {
     _delay(200);
[…]
