+++ src/StepperMotor.cpp
@@ -25,7 +25,7 @@ void StepperMotor::linkDriver(StepperDriver* _driver) {
 
 // init hardware pins
 void StepperMotor::init() {
-  if(monitor_port) monitor_port->println(F("MOT: Init"));
+  SIMPLEFOC_DEBUG("MOT: Init");
 
   // if set the phase resistance of the motor use current limit to calculate the voltage limit
   if(_isset(phase_resistance)) {
@@ -49,7 +49,7 @@ void StepperMotor::init() {
 
   _delay(500);
   // enable motor
-  if(monitor_port) monitor_port->println(F("MOT: Enable driver."));
+  SIMPLEFOC_DEBUG("MOT: Enable driver.");
   enable();
   _delay(500);
 
@@ -101,12 +101,12 @@ int  StepperMotor::initFOC( float zero_electric_offset, Direction _sensor_direct
     // added the shaft_angle update
     sensor->update();
     shaft_angle = sensor->getAngle();
-  }else if(monitor_port) monitor_port->println(F("MOT: No sensor."));
+  }else SIMPLEFOC_DEBUG("MOT: No sensor.");
 
   if(exit_flag){
-    if(monitor_port) monitor_port->println(F("MOT: Ready."));
+    SIMPLEFOC_DEBUG("MOT: Ready.");
   }else{
-    if(monitor_port) monitor_port->println(F("MOT: Init FOC failed."));
+    SIMPLEFOC_DEBUG("MOT: Init FOC failed.");
     disable();
   }
 
@@ -116,7 +116,7 @@ int  StepperMotor::initFOC( float zero_electric_offset, Direction _sensor_direct
 // Encoder alignment to electrical 0 angle
 int StepperMotor::alignSensor() {
   int exit_flag = 1; //success
-  if(monitor_port) monitor_port->println(F("MOT: Align sensor."));
+  SIMPLEFOC_DEBUG("MOT: Align sensor.");
 
   // if unknown natural direction
   if(!_isset(sensor_direction)){
@@ -147,24 +147,23 @@ int StepperMotor::alignSensor() {
     _delay(200);
     // determine the direction the sensor moved
     if (mid_angle == end_angle) {
-      if(monitor_port) monitor_port->println(F("MOT: Failed to notice movement"));
+      SIMPLEFOC_DEBUG("MOT: Failed to notice movement");
       return 0; // failed calibration
     } else if (mid_angle < end_angle) {
-      if(monitor_port) monitor_port->println(F("MOT: sensor_direction==CCW"));
+      SIMPLEFOC_DEBUG("MOT: sensor_direction==CCW");
       sensor_direction = Direction::CCW;
     } else{
-      if(monitor_port) monitor_port->println(F("MOT: sensor_direction==CW"));
+      SIMPLEFOC_DEBUG("MOT: sensor_direction==CW");
       sensor_direction = Direction::CW;
     }
     // check pole pair number
-    if(monitor_port) monitor_port->print(F("MOT: PP check: "));
     float moved =  fabs(mid_angle - end_angle);
     if( fabs(moved*pole_pairs - _2PI) > 0.5f ) { // 0.5f is arbitrary number it can be lower or higher!
-      if(monitor_port) monitor_port->print(F("fail - estimated pp:"));
-      if(monitor_port) monitor_port->println(_2PI/moved,4);
-    }else if(monitor_port) monitor_port->println(F("OK!"));
+      SIMPLEFOC_DEBUG("MOT: PP check: fail - estimated pp: ", _2PI/moved);
+    } else 
+      SIMPLEFOC_DEBUG("MOT: PP check: OK!");
 
[…]
