+++ src/common/base_classes/CurrentSense.h
@@ -13,19 +13,24 @@ class CurrentSense{
 
     /**
      *  Function intialising the CurrentSense class
-     *   All the necessary intialisations of adc and sync should be implemented here
-     */
-    virtual void init() = 0;
-    
-    /**
-     * Function intended to implement all that is needed to sync and current sensing with the driver.
-     * If no such thing is needed it can be left empty (return 1)
+     *   - All the necessary intialisations of adc and sync should be implemented here
+     *   
      * @returns -  0 - for failure &  1 - for success 
      */
-    virtual int driverSync(BLDCDriver *driver) = 0;
+    virtual int init() = 0;
+    
+    /**
+     * Linking the current sense with the motor driver
+     * Only necessary if synchronisation in between the two is required
+     */
+    void linkDriver(BLDCDriver *driver);
 
-    // calibration variables
+    // variables
     bool skip_align = false; //!< variable signaling that the phase current direction should be verified during initFOC()
+    
+    BLDCDriver* driver; //!< driver link
+    void* params = 0; //!< pointer to hardware specific parameters of current sensing
+    
     /**
      * Function intended to verify if:
      *   - phase current are oriented properly 
@@ -34,7 +39,7 @@ class CurrentSense{
      * This function corrects the alignment errors if possible ans if no such thing is needed it can be left empty (return 1)
      * @returns -  0 - for failure &  positive number (with status) - for success 
      */
-    virtual int driverAlign(BLDCDriver *driver, float voltage) = 0;
+    virtual int driverAlign(float align_voltage) = 0;
 
     /**
      *  Function rading the phase currents a, b and c
@@ -62,6 +67,8 @@ class CurrentSense{
      * @param angle_el - motor electrical angle
      */
     DQCurrent_s getFOCCurrents(float angle_el);
+
+
 };
 
 #endif
