### 2022-02-12 0ec82e3
New debug management independent of monitoring
