### 2022-02-13 4441d36
esp32 adc driver speedup
