+++ src/current_sense/hardware_specific/esp32_mcu.cpp
@@ -16,24 +16,33 @@
 #define _ADC_VOLTAGE 3.3f
 #define _ADC_RESOLUTION 4095.0f
 
+
+typedef struct ESP32MCPWMCurrentSenseParams {
+  int pins[3];
+  float adc_voltage_conv;
+  mcpwm_unit_t mcpwm_unit;
+  int buffer_index;
+} ESP32MCPWMCurrentSenseParams;
+
+
 /**
  *  Inline adc reading implementation 
 */
 // function reading an ADC value and returning the read voltage
 float _readADCVoltageInline(const int pinA, const void* cs_params){
   uint32_t raw_adc = adcRead(pinA);
-  return raw_adc * ((GenericCurrentSenseParams*)cs_params)->adc_voltage_conv;
+  return raw_adc * ((ESP32MCPWMCurrentSenseParams*)cs_params)->adc_voltage_conv;
 }
 
 // function reading an ADC value and returning the read voltage
-void* _configureADCInline(const void* driver_params, const int pinA,const int pinB,const int pinC){
+void* _configureADCInline(const void* driver_params, const int pinA, const int pinB, const int pinC){
   _UNUSED(driver_params);
 
   pinMode(pinA, INPUT);
   pinMode(pinB, INPUT);
   if( _isset(pinC) ) pinMode(pinC, INPUT);
 
-  GenericCurrentSenseParams* params = new GenericCurrentSenseParams {
+  ESP32MCPWMCurrentSenseParams* params = new ESP32MCPWMCurrentSenseParams {
     .pins = { pinA, pinB, pinC },
     .adc_voltage_conv = (_ADC_VOLTAGE)/(_ADC_RESOLUTION)
   };
@@ -46,46 +55,51 @@ void* _configureADCInline(const void* driver_params, const int pinA,const int pi
 /**
  *  Low side adc reading implementation 
 */
-static mcpwm_dev_t *MCPWM[2] = {&MCPWM0, &MCPWM1};
-int a1, a2, a3;         // Current readings from internal current sensor amplifiers
-int _pinA, _pinB, _pinC;
 
 static void IRAM_ATTR mcpwm0_isr_handler(void*);
 static void IRAM_ATTR mcpwm1_isr_handler(void*);
 byte currentState = 1;
-
-int* adc_pins;
-int* adc_buffer;
-int adc_pin_count;
-int adc_pin_read_index;
-
+// two mcpwm units 
+// - max 2 motors per mcpwm unit (6 adc channels)
+int adc_pins[2][6]={0};
+int adc_pin_count[2]={0};
+uint32_t adc_buffer[2][6]={0};
+int adc_read_index[2]={0};
 
 // function reading an ADC value and returning the read voltage
 float _readADCVoltageLowSide(const int pin, const void* cs_params){
   uint32_t raw_adc;
 
-  if (pin == _pinA) raw_adc = a1;
-  else if (pin == _pinB) raw_adc = a2;
-  else if (pin == _pinC) raw_adc = a3;
+  mcpwm_unit_t unit = ((ESP32MCPWMCurrentSenseParams*)cs_params)->mcpwm_unit;
+  int buffer_index = ((ESP32MCPWMCurrentSenseParams*)cs_params)->buffer_index;
+  float adc_voltage_conv = ((ESP32MCPWMCurrentSenseParams*)cs_params)->adc_voltage_conv;
 
-  return  raw_adc * ((GenericCurrentSenseParams*)cs_params)->adc_voltage_conv;
+  for(int i=0; i < adc_pin_count[unit]; i++){
[…]
