+++ src/current_sense/hardware_specific/samd_mcu.cpp
@@ -6,18 +6,9 @@
 #define _ADC_RESOLUTION 1024.0f
 
 // function reading an ADC value and returning the read voltage
-float _readADCVoltageInline(const int pinA, const void* cs_params){
-  digitalWrite(7,HIGH);
-  uint32_t raw_adc = analogRead(pinA);
-  digitalWrite(7,LOW);
-  return raw_adc * ((GenericCurrentSenseParams*)cs_params)->adc_voltage_conv;
-}
-
-// function reading an ADC value and returning the read voltage
 void* _configureADCInline(const void* driver_params, const int pinA,const int pinB,const int pinC){
   _UNUSED(driver_params);
-  pinMode(7,OUTPUT);
-  pinMode(pinA, INPUT);
+  pinMode(pinA, INPUT); 
   pinMode(pinB, INPUT);
   if( _isset(pinC) ) pinMode(pinC, INPUT);
 
