### 2022-02-22 92915ff
generic current sense typo
