### 2022-02-22 0ca956e
typo in action yaml
