### 2022-02-22 942b3d2
added stm32duino v2.2 to the action yaml
