### 2022-02-22 13d67f9
current sense hardware restructuring + added bg431-esc to compile check + bugfixes
