+++ src/current_sense/hardware_specific/esp32/esp32_mcu.cpp
@@ -1,6 +1,6 @@
-#include "../hardware_api.h"
-#include "../../drivers/hardware_api.h"
-#include "../../drivers/hardware_specific/esp32_driver_mcpwm.h"
+#include "../../hardware_api.h"
+#include "../../../drivers/hardware_api.h"
+#include "../../../drivers/hardware_specific/esp32_driver_mcpwm.h"
 
 #if defined(ESP_H) && defined(ARDUINO_ARCH_ESP32) && defined(SOC_MCPWM_SUPPORTED) 
 
