+++ src/current_sense/hardware_specific/stm32/b_g431/b_g431_hal.h
@@ -1,7 +1,7 @@
-#ifndef stm32g4_hal
-#define stm32g4_hal
+#ifndef B_G431_ESC1_HAL
+#define B_G431_ESC1_HAL
 
-#if defined(STM32G4xx)
+#if defined(ARDUINO_B_G431B_ESC1)
 
 #include <stm32g4xx_hal_adc.h>
 #include <stm32g4xx_hal_opamp.h>
