### 2022-02-27 cac0484
set driver initialized flag
