### 2022-03-09 0066b5e
STM32 driver refactoring
