+++ src/drivers/hardware_specific/stm32_mcu.cpp
@@ -25,6 +25,13 @@ typedef struct STM32DriverParams {
 
 
 
+#ifdef SIMPLEFOC_STM32_DEBUG
+void printTimerCombination(int numPins, PinMap* timers[], int score);
+int getTimerNumber(int timerIndex);
+#endif
+
+
+
 
 
 // setting pwm to hardware pin - instead analogWrite()
@@ -40,27 +47,24 @@ void _setPwm(HardwareTimer *HT, uint32_t channel, uint32_t value, int resolution
 }
 
 
-// init pin pwm
-HardwareTimer* _initPinPWM(uint32_t PWM_freq, int ulPin)
-{
-  PinName pin = digitalPinToPinName(ulPin);
-  TIM_TypeDef *Instance = (TIM_TypeDef *)pinmap_peripheral(pin, PinMap_PWM);
-  if (Instance == NP) {
-    Serial.print("No timer on pin ");
-    Serial.println(ulPin);
-    delay(1000);
-    return NP;
-  }
 
-  uint32_t index = get_timer_index(Instance);
+
+
+
+// init pin pwm
+HardwareTimer* _initPinPWM(uint32_t PWM_freq, PinMap* timer) {
+  // sanity check
+  if (timer==NP)
+    return NP;
+  uint32_t index = get_timer_index((TIM_TypeDef*)timer->peripheral);
   if (HardwareTimer_Handle[index] == NULL) {
-    HardwareTimer_Handle[index]->__this = new HardwareTimer((TIM_TypeDef *)pinmap_peripheral(pin, PinMap_PWM));
+    HardwareTimer_Handle[index]->__this = new HardwareTimer((TIM_TypeDef*)timer->peripheral);
     HardwareTimer_Handle[index]->handle.Init.CounterMode = TIM_COUNTERMODE_CENTERALIGNED3;
     HAL_TIM_Base_Init(&(HardwareTimer_Handle[index]->handle));
   }
   HardwareTimer *HT = (HardwareTimer *)(HardwareTimer_Handle[index]->__this);
-  uint32_t channel = STM_PIN_CHANNEL(pinmap_function(pin, PinMap_PWM));
-  HT->setMode(channel, TIMER_OUTPUT_COMPARE_PWM1, pin);
+  uint32_t channel = STM_PIN_CHANNEL(timer->function);
+  HT->setMode(channel, TIMER_OUTPUT_COMPARE_PWM1, timer->pin);
   HT->setOverflow(PWM_freq, HERTZ_FORMAT);
   HT->pause();
   HT->refresh();
@@ -68,21 +72,24 @@ HardwareTimer* _initPinPWM(uint32_t PWM_freq, int ulPin)
 }
 
 
+
+
+
+
+
 // init high side pin
-HardwareTimer* _initPinPWMHigh(uint32_t PWM_freq, int ulPin)
+HardwareTimer* _initPinPWMHigh(uint32_t PWM_freq, PinMap* timer)
 {
-  return _initPinPWM(PWM_freq, ulPin);
+  return _initPinPWM(PWM_freq, timer);
 }
 
 // init low side pin
-HardwareTimer* _initPinPWMLow(uint32_t PWM_freq, int ulPin)
+HardwareTimer* _initPinPWMLow(uint32_t PWM_freq, PinMap* timer)
 {
-  PinName pin = digitalPinToPinName(ulPin);
-  TIM_TypeDef *Instance = (TIM_TypeDef *)pinmap_peripheral(pin, PinMap_PWM);
-  uint32_t index = get_timer_index(Instance);
+  uint32_t index = get_timer_index((TIM_TypeDef*)timer->peripheral);
 
   if (HardwareTimer_Handle[index] == NULL) {
[…]
