### 2022-03-09 0051af4
stm32 driver: fix check for low-side channels
