### 2022-03-09 2d53acc
Remove string to char * warning
