### 2022-03-10 3b7b261
fix channel counting
