### 2022-03-17 b7c7136
FEAT initial low side current sensing for stm32f4
