+++ src/common/base_classes/CurrentSense.cpp
@@ -16,6 +16,16 @@ float CurrentSense::getDCCurrent(float motor_electrical_angle){
         // if only two measured currents
         i_alpha = current.a;  
         i_beta = _1_SQRT3 * current.a + _2_SQRT3 * current.b;
+    }if(!current.a){
+        // if only two measured currents
+        float a = -current.c - current.b;
+        i_alpha = a;  
+        i_beta = _1_SQRT3 * a + _2_SQRT3 * current.b;
+    }if(!current.b){
+        // if only two measured currents
+        float b = -current.a - current.c;
+        i_alpha = current.a;  
+        i_beta = _1_SQRT3 * current.a + _2_SQRT3 * b;
     }else{
         // signal filtering using identity a + b + c = 0. Assumes measurement error is normally distributed.
         float mid = (1.f/3) * (current.a + current.b + current.c);
@@ -48,6 +58,16 @@ DQCurrent_s CurrentSense::getFOCCurrents(float angle_el){
         // if only two measured currents
         i_alpha = current.a;  
         i_beta = _1_SQRT3 * current.a + _2_SQRT3 * current.b;
+    }if(!current.a){
+        // if only two measured currents
+        float a = -current.c - current.b;
+        i_alpha = a;  
+        i_beta = _1_SQRT3 * a + _2_SQRT3 * current.b;
+    }if(!current.b){
+        // if only two measured currents
+        float b = -current.a - current.c;
+        i_alpha = current.a;  
+        i_beta = _1_SQRT3 * current.a + _2_SQRT3 * b;
     } else {
         // signal filtering using identity a + b + c = 0. Assumes measurement error is normally distributed.
         float mid = (1.f/3) * (current.a + current.b + current.c);
