+++ src/common/foc_utils.h
@@ -29,6 +29,7 @@
 #define _HIGH_IMPEDANCE 0
 #define _HIGH_Z _HIGH_IMPEDANCE
 #define _ACTIVE 1
+#define _NC (NOT_SET)
 
 // dq current structure
 struct DQCurrent_s
