+++ src/current_sense/LowsideCurrentSense.cpp
@@ -43,14 +43,14 @@ void LowsideCurrentSense::calibrateOffsets(){
     // read the adc voltage 1000 times ( arbitrary number )
     for (int i = 0; i < calibration_rounds; i++) {
         _startADC3PinConversionLowSide();
-        offset_ia += _readADCVoltageLowSide(pinA, params);
-        offset_ib += _readADCVoltageLowSide(pinB, params);
+        if(_isset(pinA)) offset_ia += _readADCVoltageLowSide(pinA, params);
+        if(_isset(pinB)) offset_ib += _readADCVoltageLowSide(pinB, params);
         if(_isset(pinC)) offset_ic += _readADCVoltageLowSide(pinC, params);
         _delay(1);
     }
     // calculate the mean offsets
-    offset_ia = offset_ia / calibration_rounds;
-    offset_ib = offset_ib / calibration_rounds;
+    if(_isset(pinA)) offset_ia = offset_ia / calibration_rounds;
+    if(_isset(pinB)) offset_ib = offset_ib / calibration_rounds;
     if(_isset(pinC)) offset_ic = offset_ic / calibration_rounds;
 }
 
@@ -58,8 +58,8 @@ void LowsideCurrentSense::calibrateOffsets(){
 PhaseCurrent_s LowsideCurrentSense::getPhaseCurrents(){
     PhaseCurrent_s current;
     _startADC3PinConversionLowSide();
-    current.a = (_readADCVoltageLowSide(pinA, params) - offset_ia)*gain_a;// amps
-    current.b = (_readADCVoltageLowSide(pinB, params) - offset_ib)*gain_b;// amps
+    current.a = (!_isset(pinA)) ? 0 : (_readADCVoltageLowSide(pinA, params) - offset_ia)*gain_a;// amps
+    current.b = (!_isset(pinB)) ? 0 : (_readADCVoltageLowSide(pinB, params) - offset_ib)*gain_b;// amps
     current.c = (!_isset(pinC)) ? 0 : (_readADCVoltageLowSide(pinC, params) - offset_ic)*gain_c; // amps
     return current;
 }
