### 2022-03-17 bc2b355
FEAT current limiting in voltage mode and support for adding the KV rating of the motor
