+++ src/BLDCMotor.h
@@ -19,8 +19,9 @@ class BLDCMotor: public FOCMotor
      BLDCMotor class constructor
      @param pp pole pairs number
      @param R  motor phase resistance
+     @param KV  motor KV rating (1/K_bemf)
      */ 
-    BLDCMotor(int pp,  float R = NOT_SET);
+    BLDCMotor(int pp,  float R = NOT_SET, float KV = NOT_SET);
     
     /**
      * Function linking a motor and a foc driver 
