+++ src/common/base_classes/FOCMotor.cpp
@@ -28,6 +28,9 @@ FOCMotor::FOCMotor()
   current_sp = 0;
   current.q = 0;
   current.d = 0;
+
+  // voltage bemf 
+  voltage_bemf = 0;
   
   //monitor_port 
   monitor_port = nullptr;
@@ -105,14 +108,11 @@ void FOCMotor::monitor() {
   }
   // read currents if possible - even in voltage mode (if current_sense available)
   if(monitor_variables & _MON_CURR_Q || monitor_variables & _MON_CURR_D) {
-    DQCurrent_s c{0,0};
-    if(current_sense){
-      if(torque_controller == TorqueControlType::foc_current) c = current;
-      else{
-    	  c = current_sense->getFOCCurrents(electrical_angle);
-        c.q = LPF_current_q(c.q);
-        c.d = LPF_current_d(c.d);
-      }
+    DQCurrent_s c = current;
+    if( current_sense && torque_controller != TorqueControlType::foc_current ){
+      c = current_sense->getFOCCurrents(electrical_angle);
+      c.q = LPF_current_q(c.q);
+      c.d = LPF_current_d(c.d);
     }
     if(monitor_variables & _MON_CURR_Q) {
       monitor_port->print(c.q*1000, 2); // mAmps
