+++ src/StepperMotor.h
@@ -24,8 +24,9 @@ class StepperMotor: public FOCMotor
       StepperMotor class constructor
       @param pp  pole pair number 
       @param R  motor phase resistance
+      @param KV  motor KV rating (1/K_bemf)
     */
-    StepperMotor(int pp,  float R = NOT_SET);
+    StepperMotor(int pp,  float R = NOT_SET, float KV = NOT_SET);
 
     /**
      * Function linking a motor and a foc driver 
