### 2022-03-17 daaa0fb
bugfixes and handling only one adc nad no timer 8
