+++ src/current_sense/hardware_specific/stm32/stm32f4/stm32f4_hal.h
@@ -5,7 +5,6 @@
 
 #if defined(STM32F4xx)
 #include "stm32f4xx_hal.h"
-#include "stm32f4xx_hal_adc.h"
 #include "../../../../common/foc_utils.h"
 #include "../../../../drivers/hardware_specific/stm32_mcu.h"
 #include "../stm32_mcu.h"
