### 2022-03-20 8ddddd2
only print combination if there is one found
