### 2022-04-01 cfed655
added example sketch to find linear hall ADC center offsets, plus tweak to BLDCMotor for continuous sensor updating
