+++ src/drivers/hardware_specific/stm32_mcu.cpp
@@ -46,17 +46,19 @@ HardwareTimer* _initPinPWM(uint32_t PWM_freq, PinMap* timer) {
   if (timer==NP)
     return NP;
   uint32_t index = get_timer_index((TIM_TypeDef*)timer->peripheral);
+  bool init = false;
   if (HardwareTimer_Handle[index] == NULL) {
     HardwareTimer_Handle[index]->__this = new HardwareTimer((TIM_TypeDef*)timer->peripheral);
     HardwareTimer_Handle[index]->handle.Init.CounterMode = TIM_COUNTERMODE_CENTERALIGNED3;
     HAL_TIM_Base_Init(&(HardwareTimer_Handle[index]->handle));
+    init = true;
   }
   HardwareTimer *HT = (HardwareTimer *)(HardwareTimer_Handle[index]->__this);
   uint32_t channel = STM_PIN_CHANNEL(timer->function);
-  HT->setMode(channel, TIMER_OUTPUT_COMPARE_PWM1, timer->pin);
-  HT->setOverflow(PWM_freq, HERTZ_FORMAT);
   HT->pause();
-  HT->refresh();
+  if (init)
+    HT->setOverflow(PWM_freq, HERTZ_FORMAT);
+  HT->setMode(channel, TIMER_OUTPUT_COMPARE_PWM1, timer->pin);
   return HT;
 }
 
@@ -67,8 +69,7 @@ HardwareTimer* _initPinPWM(uint32_t PWM_freq, PinMap* timer) {
 
 
 // init high side pin
-HardwareTimer* _initPinPWMHigh(uint32_t PWM_freq, PinMap* timer)
-{
+HardwareTimer* _initPinPWMHigh(uint32_t PWM_freq, PinMap* timer) {
   return _initPinPWM(PWM_freq, timer);
 }
 
@@ -77,31 +78,39 @@ HardwareTimer* _initPinPWMLow(uint32_t PWM_freq, PinMap* timer)
 {
   uint32_t index = get_timer_index((TIM_TypeDef*)timer->peripheral);
 
+  bool init = false;
   if (HardwareTimer_Handle[index] == NULL) {
     HardwareTimer_Handle[index]->__this = new HardwareTimer((TIM_TypeDef*)timer->peripheral);
     HardwareTimer_Handle[index]->handle.Init.CounterMode = TIM_COUNTERMODE_CENTERALIGNED3;
     HAL_TIM_Base_Init(&(HardwareTimer_Handle[index]->handle));
-    TIM_OC_InitTypeDef sConfigOC = TIM_OC_InitTypeDef();
-    sConfigOC.OCMode = TIM_OCMODE_PWM2;
-    sConfigOC.Pulse = 100;
-    sConfigOC.OCPolarity = TIM_OCPOLARITY_LOW;
-    sConfigOC.OCFastMode = TIM_OCFAST_DISABLE;
-#if defined(TIM_OCIDLESTATE_SET)
-    sConfigOC.OCIdleState  = TIM_OCIDLESTATE_SET;
-#endif
-#if defined(TIM_OCNIDLESTATE_RESET)
-    sConfigOC.OCNPolarity  = TIM_OCNPOLARITY_HIGH;
-    sConfigOC.OCNIdleState = TIM_OCNIDLESTATE_RESET;
-#endif
-    uint32_t channel = STM_PIN_CHANNEL(timer->function);
-    HAL_TIM_PWM_ConfigChannel(&(HardwareTimer_Handle[index]->handle), &sConfigOC, channel);
+    init = true;
   }
   HardwareTimer *HT = (HardwareTimer *)(HardwareTimer_Handle[index]->__this);
[…]
