+++ src/StepperMotor.cpp
@@ -14,7 +14,7 @@ StepperMotor::StepperMotor(int pp, float _R, float _KV)
   // save phase resistance number
   phase_resistance = _R;
   // save back emf constant KV = 1/KV
-  K_bemf = 1.0/_KV;
+  K_bemf = _isset(_KV) ? 1.0/_KV : NOT_SET;
 
   // torque control type is voltage by default
   // current and foc_current not supported yet
