### 2022-04-06 37e1a73
initial tested low-side current sensing for f1 and f4
