+++ src/current_sense/hardware_specific/stm32/stm32f1/stm32f1_hal.cpp
@@ -1,13 +1,12 @@
-#include "stm32f4_hal.h"
+#include "stm32f1_hal.h"
 
-#if defined(STM32F4xx)
+#if defined(STM32F1xx)  
 
 #include "../../../../communication/SimpleFOCDebug.h"
 #define _TRGO_NOT_AVAILABLE 12345
 
-
 // timer to injected TRGO
-// https://github.com/stm32duino/Arduino_Core_STM32/blob/e156c32db24d69cb4818208ccc28894e2f427cfa/system/Drivers/STM32F4xx_HAL_Driver/Inc/stm32f4xx_hal_adc_ex.h#L179
+// https://github.com/stm32duino/Arduino_Core_STM32/blob/e156c32db24d69cb4818208ccc28894e2f427cfa/system/Drivers/STM32F1xx_HAL_Driver/Inc/stm32f1xx_hal_adc_ex.h#L215
 uint32_t _timerToInjectedTRGO(HardwareTimer* timer){
   if(timer->getHandle()->Instance == TIM1)  
     return ADC_EXTERNALTRIGINJECCONV_T1_TRGO;
@@ -28,17 +27,13 @@ uint32_t _timerToInjectedTRGO(HardwareTimer* timer){
 }
 
 // timer to regular TRGO
-// https://github.com/stm32duino/Arduino_Core_STM32/blob/e156c32db24d69cb4818208ccc28894e2f427cfa/system/Drivers/STM32F4xx_HAL_Driver/Inc/stm32f4xx_hal_adc.h#L331
+// https://github.com/stm32duino/Arduino_Core_STM32/blob/e156c32db24d69cb4818208ccc28894e2f427cfa/system/Drivers/STM32F1xx_HAL_Driver/Inc/stm32f1xx_hal_adc_ex.h#L215
 uint32_t _timerToRegularTRGO(HardwareTimer* timer){
-  if(timer->getHandle()->Instance == TIM2)  
-    return ADC_EXTERNALTRIGCONV_T2_TRGO;
-#ifdef TIM3 // if defined timer 3
-  else if(timer->getHandle()->Instance == TIM3) 
+  if(timer->getHandle()->Instance == TIM3) 
     return ADC_EXTERNALTRIGCONV_T3_TRGO;
-#endif
 #ifdef TIM8 // if defined timer 8
   else if(timer->getHandle()->Instance == TIM8) 
-    return ADC_EXTERNALTRIGCONV_T8_TRGO;
+    return ADC_EXTERNALTRIGINJECCONV_T8_TRGO;
 #endif
   else
     return _TRGO_NOT_AVAILABLE;
@@ -49,10 +44,10 @@ ADC_HandleTypeDef hadc;
 int _adc_init(Stm32CurrentSenseParams* cs_params, const STM32DriverParams* driver_params)
 {
   ADC_InjectionConfTypeDef sConfigInjected;
+    
   /**Configure the global features of the ADC (Clock, Resolution, Data Alignment and number of conversion) 
   */
   hadc.Instance = (ADC_TypeDef *)pinmap_peripheral(analogInputToPinName(cs_params->pins[0]), PinMap_ADC);
-
   if(hadc.Instance == ADC1) __HAL_RCC_ADC1_CLK_ENABLE();
 #ifdef ADC2  // if defined ADC2
   else if(hadc.Instance == ADC2) __HAL_RCC_ADC2_CLK_ENABLE();
@@ -66,27 +61,21 @@ int _adc_init(Stm32CurrentSenseParams* cs_params, const STM32DriverParams* drive
 #endif
     return -1; // error not a valid ADC instance
   }
-  
-  hadc.Init.ClockPrescaler = ADC_CLOCK_SYNC_PCLK_DIV4;
[…]
