### 2022-04-18 01e715d
removed old odrive example + extended simplefoc debug
