### 2022-04-18 678c578
refactored the back-emf support using KV rating of the motor - rpm/V
