+++ src/BLDCMotor.cpp
@@ -4,7 +4,7 @@
 // BLDCMotor( int pp , float R)
 // - pp            - pole pair number
 // - R             - motor phase resistance
-// - KV            - motor kv rating
+// - KV            - motor kv rating (rmp/v)
 BLDCMotor::BLDCMotor(int pp, float _R, float _KV)
 : FOCMotor()
 {
@@ -13,7 +13,7 @@ BLDCMotor::BLDCMotor(int pp, float _R, float _KV)
   // save phase resistance number
   phase_resistance = _R;
   // save back emf constant KV = 1/KV
-  K_bemf = _isset(_KV) ? 1.0/_KV : NOT_SET;
+  K_bemf = _isset(_KV) ? 1.0f/_KV/_RPM_TO_RADS : NOT_SET;
 
   // torque control type is voltage by default
   torque_controller = TorqueControlType::voltage;
