+++ src/StepperMotor.h
@@ -24,7 +24,7 @@ class StepperMotor: public FOCMotor
       StepperMotor class constructor
       @param pp  pole pair number 
       @param R  motor phase resistance
-      @param KV  motor KV rating (1/K_bemf)
+      @param KV  motor KV rating (1/K_bemf) - rpm/V
     */
     StepperMotor(int pp,  float R = NOT_SET, float KV = NOT_SET);
 
