### 2022-04-23 #4232 @sea-bass
> @sea-bass and @ixjlyons, Have you been able to get started with review? Please follow up here.

Hi @gkthiruvathukal -- I have just gone through the software for my review: https://github.com/simplefoc/Arduino-FOC/issues/174

While the software, documentation, etc. looks great, we are still unable to generate the software paper and therefore review that piece. Could you and the author take a look at this and update us when the paper is available?

Additionally, the CI pipeline includes compilation tests, but there are no unit tests, so I have recommended the addition of those for some of the core C++ source.
