### 2022-04-23 #4232 @askuric
Hey @sea-bass,
Thanks for the review!
The paper is the `joss_paper` branch [link](https://github.com/simplefoc/Arduino-FOC/tree/joss_paper/).  That was one of the suggested ways by the joss webiste ( to add a branch with the paper).

If you prefer I can add the paper to the master.
Antun
