### 2022-04-23 #4232 @sea-bass
> Hey @sea-bass, Thanks for the review! The paper is the `joss_paper` branch [link](https://github.com/simplefoc/Arduino-FOC/tree/joss_paper/). That was one of the suggested ways by the joss webiste ( to add a branch with the paper).
>
> If you prefer I can add the paper to the master. Antun

Thanks, @askuric -- turns out the main issue description didn't specify a branch, so it defaulted. I've just edited it to the `joss_paper` branch. Let's see how this turns out.
