### 2022-04-25 249e6a3
enabled support for _NC in current sense definition
