+++ src/current_sense/hardware_specific/due_mcu.cpp
@@ -10,8 +10,8 @@
 void* _configureADCInline(const void* driver_params, const int pinA,const int pinB,const int pinC){
   _UNUSED(driver_params);
 
-  pinMode(pinA, INPUT);
-  pinMode(pinB, INPUT);
+  if( _isset(pinA) ) pinMode(pinA, INPUT);
+  if( _isset(pinB) ) pinMode(pinB, INPUT);
   if( _isset(pinC) ) pinMode(pinC, INPUT);
 
   GenericCurrentSenseParams* params = new GenericCurrentSenseParams {
