### 2022-04-26 daaa3e1
removed K_bemf and added KV_rating
