### 2022-04-28 #4232 @ixjlyons
My review:

----

SimpleFOC is a library offering a modular approach to implementing field oriented control with various hardware components. I'm currently not able to directly verify functionality of the library, but I've been able to successfully compile several examples and have reviewed portions of the code. The code itself, along with the very well-done documentation, appear to be a great resource for implementing advanced motor control for a variety of applications. I recommend accepting.

### General Comments

I may have missed it, but I haven't come across clear community guidelines as required by JOSS in the documentation or README. I think this is the only blocking issue. Aside from that, I only have a few minor comments:

- It took me some time to find the API docs - perhaps there could be a direct link in the sidebar?
- In the [Writing the Code](https://docs.simplefoc.com/code#step-1-position-sensor-setup) page, the buttons to toggle position sensor choices (magnetic sensor vs. encoder) don't seem to work.

### Paper

No major concerns with the paper. Below are some minor comments:

- Line 10: haven't seen the use of "alternative current" before, consider revising to "alternating current". Also consider changing to lower case "alternating current" and "brushless"
- Line 16: looks like there are extra spaces inside the parentheses around "see Figure 1"
- Line 19: it seems like the FOC *approach* is general, while the *implementation* is hardware-specific
- Line 43: should probably use singular "motor" for consistency with other items in the list, or rework to "...supports many combinations of motors, sensors, ..."
- Line 48: lowercase "stepper" except in the class names
- Line 53: maybe replace comma with a semicolon
- Line 75: consider revising to "...configurability of SimpleFOC..." (i.e. remove "the"), or maybe "...the SimpleFOC library..."
- Line 83, should have a comma after "bandwidth"
