### 2022-05-01 bc0f5a4
current sense better align + example refactoring
