+++ src/current_sense/InlineCurrentSense.cpp
@@ -69,95 +69,130 @@ PhaseCurrent_s InlineCurrentSense::getPhaseCurrents(){
 // 3 - success but gains inverted
 // 4 - success but pins reconfigured and gains inverted
 int InlineCurrentSense::driverAlign(float voltage){
+    
     int exit_flag = 1;
     if(skip_align) return exit_flag;
 
-    // set phase A active and phases B and C down
-    driver->setPwm(voltage, 0, 0);
-    _delay(200);
-    PhaseCurrent_s c = getPhaseCurrents();
-    // read the current 100 times ( arbitrary number )
-    for (int i = 0; i < 100; i++) {
-        PhaseCurrent_s c1 = getPhaseCurrents();
-        c.a = c.a*0.6f + 0.4f*c1.a;
-        c.b = c.b*0.6f + 0.4f*c1.b;
-        c.c = c.c*0.6f + 0.4f*c1.c;
-        _delay(3);
-    }
-    driver->setPwm(0, 0, 0);
-    // align phase A
-    float ab_ratio = fabs(c.a / c.b);
-    float ac_ratio = c.c ? fabs(c.a / c.c) : 0;
-    if( ab_ratio > 1.5f ){ // should be ~2
-        gain_a *= _sign(c.a);
-    }else if( ab_ratio < 0.7f ){ // should be ~0.5
-        // switch phase A and B
-        int tmp_pinA = pinA;
-        pinA = pinB;
-        pinB = tmp_pinA;
-        gain_a *= _sign(c.b);
-        exit_flag = 2; // signal that pins have been switched
-    }else if(_isset(pinC) &&  ac_ratio < 0.7f ){ // should be ~0.5
-        // switch phase A and C
-        int tmp_pinA = pinA;
-        pinA = pinC;
-        pinC= tmp_pinA;
-        gain_a *= _sign(c.c);
-        exit_flag = 2;// signal that pins have been switched
-    }else{
-        // error in current sense - phase either not measured or bad connection
-        return 0;
+    if(_isset(pinA)){
+        // set phase A active and phases B and C down
+        driver->setPwm(voltage, 0, 0);
+        _delay(2000);
+        PhaseCurrent_s c = getPhaseCurrents();
+        // read the current 100 times ( arbitrary number )
+        for (int i = 0; i < 100; i++) {
+            PhaseCurrent_s c1 = getPhaseCurrents();
+            c.a = c.a*0.6f + 0.4f*c1.a;
+            c.b = c.b*0.6f + 0.4f*c1.b;
+            c.c = c.c*0.6f + 0.4f*c1.c;
+            _delay(3);
+        }
+        driver->setPwm(0, 0, 0);
+        // align phase A
+        float ab_ratio = c.b ? fabs(c.a / c.b) : 0;
+        float ac_ratio = c.c ? fabs(c.a / c.c) : 0;
+        if(_isset(pinB) && ab_ratio > 1.5f ){ // should be ~2
+            gain_a *= _sign(c.a);
+        }else if(_isset(pinC) && ac_ratio > 1.5f ){ // should be ~2
+            gain_a *= _sign(c.a);
+        }else if(_isset(pinB) && ab_ratio < 0.7f ){ // should be ~0.5
+            // switch phase A and B
+            int tmp_pinA = pinA;
+            pinA = pinB;
+            pinB = tmp_pinA;
+            gain_a *= _sign(c.b);
+            exit_flag = 2; // signal that pins have been switched
+        }else if(_isset(pinC) &&  ac_ratio < 0.7f ){ // should be ~0.5
+            // switch phase A and C
+            int tmp_pinA = pinA;
+            pinA = pinC;
+            pinC= tmp_pinA;
+            gain_a *= _sign(c.c);
+            exit_flag = 2;// signal that pins have been switched
[…]
