### 2022-05-04 40e1d21
change StepDirListener to use RISING interrupts
