+++ src/communication/StepDirListener.h
@@ -47,11 +47,13 @@ class StepDirListener
     int pin_step; //!< step pin
     int pin_dir; //!< direction pin
     long count; //!< current counter value - should be set to 0 for homing
+    PinStatus polarity = RISING; //!< polarity of the step pin
 
   private:
     float* attached_variable = nullptr; //!< pointer to the attached variable 
     float counter_to_value; //!< step counter to value 
-    bool step_active = 0; //!< current step pin status (HIGH/LOW) - debouncing variable
+    //bool step_active = 0; //!< current step pin status (HIGH/LOW) - debouncing variable
+
 };
 
 #endif
