### 2022-05-06 #4232 @askuric
Hey @ixjlyons and @sea-bass,
Thank you very much for your insightful and supporting reviews!

@sea-bass
With the new release of the library we've spent some time on the gifs/moving images and made them a lot more clear. Thank you for the suggestion.
We'll also consider adding more functionality unitary testing, but for now as the code is intended to be run on different MCU architectures and interface with different hardware components the full scale testing is a bit hard to do. But as you well pointed out there are ways to do it and we should do more in this regard. :D

@ixjlyons
We've updated the paper to account for your remarks.
In the docs, the button has been fixed and we've added much more examples which can be interchanged with a click of the button to make it easier to understand.

Regarding the community guidelines.
As many of our users are not really git users we have opted to the scheme where they are free to propose and share their problems/contributions within the community framework And if they are motivated we help them to create the PR request and we really treat each one case by case.
But you are absolutely right that we would need a bit more structures issue template and PR template in order to make contribution and issue reporting more efficient and more clear.
We add the guidelines to the github repo this till the end of this week.

Thanks again for your time and let me know if there is something else that you'd like us to consider!
