### 2022-05-08 #4232 @ixjlyons
> @ixjlyons We've updated the paper to account for your remarks. In the docs, the button has been fixed and we've added much more examples which can be interchanged with a click of the button to make it easier to understand.

Looks good to me :+1:

> Regarding the community guidelines. As many of our users are not really git users we have opted to the scheme where they are free to propose and share their problems/contributions within the community framework And if they are motivated we help them to create the PR request and we really treat each one case by case.

I think that's fine -- my understanding is there just needs to be [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) written somewhere to make expectations clear. I'd say the links in the header of the docs cover getting help and maybe even bringing up issues. "Please open an issue to propose improvements" in the README would be sufficient I think. Pull request / issue templates are great additions too.
