### 2022-05-09 #4232 @sea-bass
> @sea-bass With the new release of the library we've spent some time on the gifs/moving images and made them a lot more clear. Thank you for the suggestion. We'll also consider adding more functionality unitary testing, but for now as the code is intended to be run on different MCU architectures and interface with different hardware components the full scale testing is a bit hard to do. But as you well pointed out there are ways to do it and we should do more in this regard. :D

Of course! I was only suggesting unit testing, as having full-scale (or system-level) testing for a hardware-focused library would be unreasonable.
