### 2022-05-15 #4232 @askuric
Hey @ixjlyons,
Ok so we've added a short contributing webpage as the addition to the contact page in the [docs](https://docs.simplefoc.com/).

Also the [README](https://github.com/simplefoc/Arduino-FOC) was extended with the section about contributing and different community platforms that are available.

Finally, I've added the first version of issue templates that should enable users to post more direct and precise questions/issues. For the PRs it is a bit more tricky so we will provide the templates a bit later.
In any case, the docs and the README and our communications are clearly oriented in a way to encourage the users to contribute in any way they feel most conformable about.
