### 2022-05-25 #4232 @sea-bass
According to the guidelines, the requirements are:

**Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

If there is no plan to create automated unit tests / a CI pipeline for this review process (which is a totally fair decision), could you point me to the "manual steps" that would meet the criteria above? And if not available, could some documentation on testing this software and verifying it works as intended be made?

Thanks!
