+++ src/communication/StepDirListener.h
@@ -47,7 +47,7 @@ class StepDirListener
     int pin_step; //!< step pin
     int pin_dir; //!< direction pin
     long count; //!< current counter value - should be set to 0 for homing
-    PinStatus polarity = RISING; //!< polarity of the step pin
+    int polarity = RISING; //!< polarity of the step pin
 
   private:
     float* attached_variable = nullptr; //!< pointer to the attached variable 
