### 2022-06-10 #4232 @askuric
Hey everyone,

I am sorry for a late reply!

At the moment the CI pipeline for unit testing of the software components is really not our priority. Mostly because the most of the problems due to the library code comes from the MCU specific issues which cannot really be unit tested.
We are definitely interested in integrating the unit testing at some point (the sooner the better) but at the moment we do not have the time resources to do it :D
We are however making sure that the library compiles for each one of the supported platforms using the githubs CIs, but for the moment not more than that.

However, we do have guidelines in the docs that explain how each user can verify that each part of the library works correctly with their hardware. This is a very important step for setting up any of the users hardware setups, regardless of how standard or nonstandard they are. So in an way each user really needs to test and verify that the library code works well with their hardware. Additionally we support many different debugging and monitoring features which facilitate testing as well.
Getting started explains step by step testing/verification procedure: [docs page](https://docs.simplefoc.com/example_from_scratch).
Additionally there are many examples codes in the library that are intended for testing and verifying that library works as expected with given hardware.

As basically all of our users use our code only in combination with hardware components, till this point we did not consider putting in place any purely software testing procedures, all of our testing procedures are coupled with hardware components.
