### 2022-06-10 #4232 @sea-bass
@askuric thank you for clarifying that and pointing me to the right resources -- I think that absolutely meets the testing requirements, and agree with your prioritization especially given the hardware-focused nature of the project.

LGTM :)
