### 2022-06-10 #4232 @askuric
Thanks @ixjlyons  and @sea-bass  for yout reviews, we appreciate it!
@gkthiruvathukal  don't hesitate to let us know what we need to do to advance to the next step.

I have one question though, would it be possible to chnage the title of the paper to:
"SimpleFOC: A Cross-Platform Field Oriented Control (FOC) Library for Controlling Brushless Direct Current (BLDC) and Stepper Motors"

I think it might be important to add 'Cross-Platform' to the title as one of out main contributions.
