### 2022-06-13 #4232 @askuric
Hello @gkthiruvathukal,

1. We already have a zenodo entry for our releases so I've updated the metadata of the paper for our latest release. I hope that is fine with you.
> This release does not have paper source inside, let me know if we are required integrate it into our release.

2. I've changed the title of the paper just a bit, by adding "Cross-Platform" to the title
"SimpleFOC: A **Cross-Platform** Field Oriented Control (FOC) Library for Controlling Brushless Direct Current (BLDC) and Stepper Motors"
I hope that you are fine with it, as it is one of our main contributions. It is updated in paper source in the the joss_paper branch and in the Zenodo entry.
> If the title change is not possible I'll revert to the original version as soon as possible.

So our software release version is:
v2.2.2

The Zenodo DOI value:
10.5281/zenodo.6510536
