### 2022-06-14 4247ac2
leonardo small memory issues
