### 2022-06-14 c7d2a20
leonardo and mega2560 #190 compile error + added the autocompile to the CI
