### 2022-06-23 #4232 @askuric
@gkthiruvathukal  thanks!
The book Méchatronique does not have a DOI number, the suggested DOI is not referencing the same book.
In the paper we have provided the ISBN number of the book and the year of publishing.
