### 2022-06-24 5ea8a06
Add Linear Hall sensor .
添加线性霍尔传感器模块。
两个线性霍尔90度安装检测一对极的径向磁铁来获取转子角度。
一个线性霍尔产生sin线性波形
另一个线性霍尔产生cos线性波形
滤波后用反正切函数(atan2)计算出转子角度。
