### 2022-06-25 #4232 @Kevin-Mattheus-Moerman
@askuric thanks. Can you please group the `community member` affiliations into 1 and remove the countries there? This is not needed here. Thanks.
